requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
httpx==0.28.1
fastapi==0.109.0
uvicorn==0.27.0
psutil==5.9.8
mcrcon==0.7.0